+++ src/main/java/grepp/NBE5_6_2_Team03/api/controller/exchange/ExchangeController.java
@@ -3,7 +3,6 @@ package grepp.NBE5_6_2_Team03.api.controller.exchange;
 import grepp.NBE5_6_2_Team03.api.controller.exchange.dto.ExchangeDto;
 import grepp.NBE5_6_2_Team03.domain.exchange.entity.ExchangeRateEntity;
 import grepp.NBE5_6_2_Team03.domain.exchange.repository.ExchangeRateRepository;
-import grepp.NBE5_6_2_Team03.global.mapper.ExchangeMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/exchange")
@@ -23,12 +21,10 @@ public class ExchangeController {
 
     @GetMapping("/latest")
     public ExchangeDto getLatest(@RequestParam String code){
-        ExchangeRateEntity entity = exchangeRateRepository.findTopByCurUnitOrderByDateDesc(code)
+        ExchangeRateEntity entity = exchangeRateRepository.findLatestByCurUnit(code)
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 
-        log.info("{}", entity);
-
-        return ExchangeMapper.toDto(entity);
+        return ExchangeDto.toDto(entity);
     }
 
 }
