+++ src/main/java/grepp/NBE5_6_2_Team03/domain/exchange/entity/ExchangeRateEntity.java
@@ -9,15 +9,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 
-@Entity
 @Table(name = "exchange_rate")
-@Getter
+@Getter @Entity
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-@ToString
 public class ExchangeRateEntity {
 
     @Id
